refactor: adjust code structure and imports
