+++ src/tests/unit/create-expense.test.ts
@@ -1,6 +1,6 @@
-import { beforeEach, describe, expect, it } from 'vitest';
 import { InMemoryExpensesRepository } from '@/repositories/in-memory/in-memory-expenses-repository';
-import { CreateExpenseUseCase, CreateExpenseUseCaseRequest } from '@/use-cases/create-expense';
+import { CreateExpenseUseCase, CreateExpenseUseCaseRequest } from '@/use-cases/expense-use-cases/create-expense';
+import { beforeEach, describe, expect, it } from 'vitest';
 
 let expensesRepository: InMemoryExpensesRepository;
 let sut: CreateExpenseUseCase;
@@ -26,7 +26,7 @@ describe('Create Expense Use Case', () => {
     expect(expense.id).toEqual(expect.any(Number));
     expect(expense.description).toBe(expenseData.description);
     expect(expense.date).toEqual(expenseData.date);
-    expect(expense.value.toNumber()).toBe(expenseData.value); 
+    expect(expense.value.toNumber()).toBe(expenseData.value);
     expect(expense.categoryId).toBe(expenseData.categoryId);
     expect(expense.payment_method).toBe(expenseData.payment_method);
     expect(expense.userId).toBe(expenseData.userId);
