+++ src/tests/unit/get-expense-details.test.ts
@@ -1,6 +1,6 @@
-import { beforeEach, describe, expect, it } from 'vitest';
 import { InMemoryExpensesRepository } from '@/repositories/in-memory/in-memory-expenses-repository';
-import { GetExpenseDetailsUseCase } from '@/use-cases/get-expense-details';
+import { GetExpenseDetailsUseCase } from '@/use-cases/expense-use-cases/get-expense-details';
+import { beforeEach, describe, expect, it } from 'vitest';
 
 let expensesRepository: InMemoryExpensesRepository;
 let sut: GetExpenseDetailsUseCase;
