+++ src/tests/unit/get-filtered-expenses.test.ts
@@ -1,6 +1,6 @@
-import { beforeEach, describe, expect, it } from 'vitest';
 import { InMemoryExpensesRepository } from '@/repositories/in-memory/in-memory-expenses-repository';
-import { GetFilteredExpensesUseCase } from '@/use-cases/get-filtered-expenses';
+import { GetFilteredExpensesUseCase } from '@/use-cases/expense-use-cases/get-filtered-expenses';
+import { beforeEach, describe, expect, it } from 'vitest';
 
 let expensesRepository: InMemoryExpensesRepository;
 let sut: GetFilteredExpensesUseCase;
