+++ src/tests/unit/register.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { InMemoryUsersRepository } from '@/repositories/in-memory/in-memory-users-repository';
-import { RegisterUseCase } from '@/use-cases/register';
+import { RegisterUseCase } from '@/use-cases/user-use-cases/register';
 import { compare } from 'bcryptjs';
 import { UserAlreadyExistsError } from '@/use-cases/errors/user-already-exists-error';
