+++ src/use-cases/expense-use-cases/get-all-expenses.ts
