+++ src/use-cases/factories/make-authenticate-use-case.ts
@@ -1,4 +1,4 @@
-import { AuthenticateUseCase } from "../authenticate";
+import { AuthenticateUseCase } from "../user-use-cases/authenticate";
 
 export function makeAuthenticateUseCase() {
     const authenticateUseCase = new AuthenticateUseCase();
