+++ src/use-cases/factories/make-create-expense-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaExpensesRepository } from "@/repositories/prisma/prisma-expenses-repository";
-import { CreateExpenseUseCase } from "../create-expense";
+import { CreateExpenseUseCase } from "../expense-use-cases/create-expense";
 
 export function makeCreateExpenseUseCase() {
   const expensesRepository = new PrismaExpensesRepository();
