+++ src/use-cases/factories/make-forgot-password-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaUsersRepository } from "@/repositories/prisma/prisma-users-repository";
-import { ForgotPasswordUseCase } from "../forgot-password";
+import { ForgotPasswordUseCase } from "../user-use-cases/forgot-password";
 
 export function makeForgotPasswordUseCase() {
     const usersRepository = new PrismaUsersRepository()
