+++ src/use-cases/factories/make-get-all-expenses-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaExpensesRepository } from "@/repositories/prisma/prisma-expenses-repository";
-import { GetAllExpensesUseCase } from "../get-all-expenses";
+import { GetAllExpensesUseCase } from "../expense-use-cases/get-all-expenses";
 
 export function makeGetAllExpensesUseCase() {
   const expensesRepository = new PrismaExpensesRepository();
