+++ src/use-cases/factories/make-get-expense-details-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaExpensesRepository } from "@/repositories/prisma/prisma-expenses-repository";
-import { GetExpenseDetailsUseCase } from "../get-expense-details";
+import { GetExpenseDetailsUseCase } from "../expense-use-cases/get-expense-details";
 
 export function makeGetExpenseDetailsUseCase() {
   const expensesRepository = new PrismaExpensesRepository();
