+++ src/use-cases/factories/make-register-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaUsersRepository } from "@/repositories/prisma/prisma-users-repository";
-import { RegisterUseCase } from "../register";
+import { RegisterUseCase } from "../user-use-cases/register";
 
 export function makeRegisterUseCase() {
     const usersRepository = new PrismaUsersRepository()
