+++ src/use-cases/factories/make-update-expense-use-case.ts
@@ -1,8 +1,8 @@
 import { PrismaExpensesRepository } from "@/repositories/prisma/prisma-expenses-repository";
-import { UpdateExpenseUseCase } from "../update-expense";
+import { UpdateExpenseUseCase } from "../expense-use-cases/update-expense";
 
 export function makeUpdateExpenseUseCase() {
-  const expensesRepository = new PrismaExpensesRepository();
-  const updateExpenseUseCase = new UpdateExpenseUseCase(expensesRepository);
-  return updateExpenseUseCase;
+    const expensesRepository = new PrismaExpensesRepository();
+    const updateExpenseUseCase = new UpdateExpenseUseCase(expensesRepository);
+    return updateExpenseUseCase;
 }
