+++ src/use-cases/factories/make-update-user-profile-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaUsersRepository } from "@/repositories/prisma/prisma-users-repository";
-import { UpdateProfileUseCase } from "../update-user-profile";
+import { UpdateProfileUseCase } from "../user-use-cases/update-user-profile";
 
 export function makeUpdateProfileUseCase() {
     const usersRepository = new PrismaUsersRepository();
